AI Server/requirements.txt
Flask
celery
redis
requests
gdown
moviepy
faster_whisper
python-dotenv
